services/api/src/security/jwt.py
from datetime import datetime, timedelta, timezone
from jose import jwt
from src.config import JWT_SECRET, JWT_ALGORITHM, JWT_EXPIRES_MINUTES


def create_token(data: dict) -> str:
    payload = data.copy()
    payload["exp"] = datetime.now(timezone.utc) + timedelta(minutes=JWT_EXPIRES_MINUTES)
    return jwt.encode(payload, JWT_SECRET, algorithm=JWT_ALGORITHM)


def decode_token(token: str):
    try:
        return jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
    except Exception:
        return None
