pyproject.toml
[project]
name = "jamble-shopping-inventory-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.5",
    "functions-framework>=3.10.0",
    "google-cloud-firestore>=2.23.0",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "functions-framework>=3.10.0",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning:google.auth._default",
]
